[refactor] : 상품 등록 시 front에서 validation

+++ src/main/java/shop/itbook/itbookfront/elastic/dto/response/ProductSampleResponseDto.java
@@ -15,7 +15,7 @@ public class ProductSampleResponseDto {
     private String thumbnailUrl;
     private Long fixedPrice;
     private Integer increasePointPercent;
-    private Integer discountPercent;
+    private Double discountPercent;
     private Long rawPrice;
     private Boolean isForceSoldOut;
     private Integer stock;
