+++ src/main/java/shop/itbook/itbookfront/product/dto/response/ProductDetailsResponseDto.java
@@ -41,7 +41,7 @@ public class ProductDetailsResponseDto {
 
     private Long fixedPrice;
 
-    private Integer discountPercent;
+    private Double discountPercent;
 
     private String fileThumbnailsUrl;
